DownloadStatusResponseBody#downloadedFiles 타입을 long으로 변경

+++ src/main/java/jp/co/canon/cks/eec/fs/rssportal/model/DownloadStatusResponseBody.java
@@ -13,7 +13,7 @@ public class DownloadStatusResponseBody {
     private String downloadId;
     private String status = "invalid-id";
     private int totalFiles = -1;
-    private int downloadedFiles = -1;
+    private long downloadedFiles = -1;
     private String downloadUrl = "";
 
     public DownloadStatusResponseBody(@NonNull FileDownloader fileDownloader, @NonNull final String dlId) {
